[Decomp][Spell] Cohesion split: extract object/combat handlers to SpellEffectObjectCombat.cpp
Move the remaining EffectThreat..EffectPlayerPull block (threat, summon-object,
duel, activate-object incl. the generateActivateCommand file-static, glyph,
enchant, pet, resurrect, charge/leap, taxi) out of SpellEffects.cpp into a new
SpellEffectObjectCombat.cpp translation unit.

This finishes the SpellEffects.cpp decomposition: the file is now just the
license/includes preamble, the SpellEffects[] dispatch table, and the three
EffectEmpty/NULL/Unused no-op stubs (~290 lines, down from 14,579). All handler
bodies live in themed SpellEffectXxx.cpp TUs; the table resolves them by address
at link time. Byte-identical moves, no behaviour change.
