+++ src/game/WorldHandlers/SpellEffects.cpp
@@ -288,2531 +288,3 @@ void Spell::EffectUnused(SpellEffectEntry const* /*effect*/)
 {
     // NOT USED BY ANY SPELL OR USELESS OR IMPLEMENTED IN DIFFERENT WAY IN MANGOS
 }
-
-/**
- * @brief Opens or sends loot for the specified object guid.
- *
- * @param guid The loot source guid.
- * @param loottype The loot type to open.
- * @param lockType The lock interaction type.
- */
-void Spell::SendLoot(ObjectGuid guid, LootType loottype, LockType lockType)
-{
-    if (gameObjTarget)
-    {
-        switch (gameObjTarget->GetGoType())
-        {
-            case GAMEOBJECT_TYPE_DOOR:
-            case GAMEOBJECT_TYPE_BUTTON:
-            case GAMEOBJECT_TYPE_QUESTGIVER:
-            case GAMEOBJECT_TYPE_SPELL_FOCUS:
-            case GAMEOBJECT_TYPE_GOOBER:
-                gameObjTarget->Use(m_caster);
-                return;
-
-            case GAMEOBJECT_TYPE_CHEST:
-                gameObjTarget->Use(m_caster);
-                // Don't return, let loots been taken
-                break;
-
-            case GAMEOBJECT_TYPE_TRAP:
-                if (lockType == LOCKTYPE_DISARM_TRAP)
-                {
-                    gameObjTarget->SetLootState(GO_JUST_DEACTIVATED);
-                    return;
-                }
-                sLog.outError("Spell::SendLoot unhandled locktype %u for GameObject trap (entry %u) for spell %u.", lockType, gameObjTarget->GetEntry(), m_spellInfo->Id);
-                return;
-            default:
-                sLog.outError("Spell::SendLoot unhandled GameObject type %u (entry %u) for spell %u.", gameObjTarget->GetGoType(), gameObjTarget->GetEntry(), m_spellInfo->Id);
-                return;
-        }
-    }
-
-    if (m_caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    // Send loot
-    ((Player*)m_caster)->SendLoot(guid, loottype);
-}
-
-/**
- * @brief Opens a locked game object or item and awards related skill progress.
- *
- * @param effect The open-lock effect index.
- */
-void Spell::EffectOpenLock(SpellEffectEntry const* effect)
-{
-    if (!m_caster || m_caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        DEBUG_LOG("WORLD: Open Lock - No Player Caster!");
-        return;
-    }
-
-    Player* player = (Player*)m_caster;
-
-    uint32 lockId = 0;
-    ObjectGuid guid;
-
-    // Get lockId
-    if (gameObjTarget)
-    {
-        GameObjectInfo const* goInfo = gameObjTarget->GetGOInfo();
-        // Arathi Basin banner opening !
-        if ((goInfo->type == GAMEOBJECT_TYPE_BUTTON && goInfo->button.noDamageImmune) ||
-                (goInfo->type == GAMEOBJECT_TYPE_GOOBER && goInfo->goober.losOK))
-        {
-            // CanUseBattleGroundObject() already called in CheckCast()
-            // in battleground check
-            if (BattleGround* bg = player->GetBattleGround())
-            {
-                // check if it's correct bg
-                if (bg->GetTypeID() == BATTLEGROUND_AB || bg->GetTypeID() == BATTLEGROUND_AV)
-                {
-                    bg->EventPlayerClickedOnFlag(player, gameObjTarget);
-                }
-                return;
-            }
-        }
-        else if (goInfo->type == GAMEOBJECT_TYPE_FLAGSTAND)
-        {
-            // CanUseBattleGroundObject() already called in CheckCast()
-            // in battleground check
-            if (BattleGround* bg = player->GetBattleGround())
-            {
-                if (bg->GetTypeID() == BATTLEGROUND_EY)
-                {
-                    bg->EventPlayerClickedOnFlag(player, gameObjTarget);
-                }
-                return;
-            }
-        }
-        lockId = goInfo->GetLockId();
-        guid = gameObjTarget->GetObjectGuid();
-    }
-    else if (itemTarget)
-    {
-        lockId = itemTarget->GetProto()->LockID;
-        guid = itemTarget->GetObjectGuid();
-    }
-    else
-    {
-        DEBUG_LOG("WORLD: Open Lock - No GameObject/Item Target!");
-        return;
-    }
-
-    SkillType skillId = SKILL_NONE;
-    int32 reqSkillValue = 0;
-    int32 skillValue;
-
-    SpellCastResult res = CanOpenLock(SpellEffectIndex(effect->EffectIndex), lockId, skillId, reqSkillValue, skillValue);
-    if (res != SPELL_CAST_OK)
-    {
-        SendCastResult(res);
-        return;
-    }
-
-    // mark item as unlocked
-    if (itemTarget)
-    {
-        itemTarget->SetFlag(ITEM_FIELD_FLAGS, ITEM_DYNFLAG_UNLOCKED);
-    }
-
-    SendLoot(guid, LOOT_SKINNING, LockType(effect->EffectMiscValue));
-
-    // not allow use skill grow at item base open
-    if (!m_CastItem && skillId != SKILL_NONE)
-    {
-        // update skill if really known
-        if (uint32 pureSkillValue = player->GetPureSkillValue(skillId))
-        {
-            if (gameObjTarget)
-            {
-                // Allow one skill-up until respawned
-                if (!gameObjTarget->IsInSkillupList(player) &&
-                        player->UpdateGatherSkill(skillId, pureSkillValue, reqSkillValue))
-                    gameObjTarget->AddToSkillupList(player);
-            }
-            else if (itemTarget)
-            {
-                // Do one skill-up
-                player->UpdateGatherSkill(skillId, pureSkillValue, reqSkillValue);
-            }
-        }
-    }
-}
-
-/**
- * @brief Replaces the cast item with another item entry.
- *
- * @param effect The effect index defining the replacement item.
- */
-void Spell::EffectSummonChangeItem(SpellEffectEntry const* effect)
-{
-    if (m_caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Player* player = (Player*)m_caster;
-
-    // applied only to using item
-    if (!m_CastItem)
-    {
-        return;
-    }
-
-    // ... only to item in own inventory/bank/equip_slot
-    if (m_CastItem->GetOwnerGuid() != player->GetObjectGuid())
-    {
-        return;
-    }
-
-    uint32 newitemid = effect->EffectItemType;
-    if (!newitemid)
-    {
-        return;
-    }
-
-    Item* oldItem = m_CastItem;
-
-    // prevent crash at access and unexpected charges counting with item update queue corrupt
-    ClearCastItem();
-
-    player->ConvertItem(oldItem, newitemid);
-}
-
-/**
- * @brief Grants weapon or armor proficiency to a player target.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectProficiency(SpellEffectEntry const* /*effect*/)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-    Player *p_target = (Player*)unitTarget;
-
-    SpellEquippedItemsEntry const* eqItems = m_spellInfo->GetSpellEquippedItems();
-
-    if (!eqItems)
-    {
-        return;
-    }
-
-    if (eqItems->EquippedItemClass == ITEM_CLASS_WEAPON && !(p_target->GetWeaponProficiency() & eqItems->EquippedItemSubClassMask))
-    {
-        p_target->AddWeaponProficiency(eqItems->EquippedItemSubClassMask);
-        p_target->SendProficiency(ITEM_CLASS_WEAPON, p_target->GetWeaponProficiency());
-    }
-    if (eqItems->EquippedItemClass == ITEM_CLASS_ARMOR && !(p_target->GetArmorProficiency() & eqItems->EquippedItemSubClassMask))
-    {
-        p_target->AddArmorProficiency(eqItems->EquippedItemSubClassMask);
-        p_target->SendProficiency(ITEM_CLASS_ARMOR, p_target->GetArmorProficiency());
-    }
-}
-
-/**
- * @brief Creates and attaches an area aura for the current unit target.
- *
- * @param effect The area aura effect index.
- */
-void Spell::EffectApplyAreaAura(SpellEffectEntry const* effect)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-    if (!unitTarget->IsAlive())
-    {
-        return;
-    }
-
-    AreaAura* Aur = new AreaAura(m_spellInfo, SpellEffectIndex(effect->EffectIndex), &m_currentBasePoints[effect->EffectIndex], m_spellAuraHolder, unitTarget, m_caster, m_CastItem);
-    m_spellAuraHolder->AddAura(Aur, SpellEffectIndex(effect->EffectIndex));
-}
-
-void Spell::EffectSummonType(SpellEffectEntry const* effect)
-{
-    // if this spell already have an aura applied cancel the summon
-    if (m_caster->HasAura(m_spellInfo->Id))
-    {
-        return;
-    }
-
-    uint32 prop_id = effect->EffectMiscValueB;
-    SummonPropertiesEntry const *summon_prop = sSummonPropertiesStore.LookupEntry(prop_id);
-    if (!summon_prop)
-    {
-        sLog.outError("EffectSummonType: Unhandled summon type %u", prop_id);
-        return;
-    }
-
-    // Pet's are atm handled differently
-    if (summon_prop->Group == SUMMON_PROP_GROUP_PETS && prop_id != 1562)
-    {
-        DoSummonPet(effect);
-    }
-
-    // Expected Amount: TODO - there are quite some exceptions (like totems, engineering dragonlings..)
-    uint32 amount = damage > 0 ? damage : 1;
-
-    // basepoints of SUMMON_PROP_GROUP_VEHICLE is often a spellId, set amount to 1
-    if ((summon_prop->Group == SUMMON_PROP_GROUP_VEHICLE || (prop_id == 1961)) || summon_prop->Group == SUMMON_PROP_GROUP_UNCONTROLLABLE_VEHICLE || summon_prop->Group == SUMMON_PROP_GROUP_CONTROLLABLE)
-    {
-        amount = 1;
-    }
-
-    // Get casting object
-    WorldObject* realCaster = GetCastingObject();
-    if (!realCaster)
-    {
-        sLog.outError("EffectSummonType: No Casting Object found for spell %u, (caster = %s)", m_spellInfo->Id, m_caster->GetGuidStr().c_str());
-        return;
-    }
-
-    Unit* responsibleCaster = m_originalCaster;
-    if (realCaster->GetTypeId() == TYPEID_GAMEOBJECT)
-    {
-        responsibleCaster = ((GameObject*)realCaster)->GetOwner();
-    }
-
-    // Expected Level                                       (Totem, Pet and Critter may not use this)
-    uint32 level = responsibleCaster ? responsibleCaster->getLevel() : m_caster->getLevel();
-    // level of creature summoned using engineering item based at engineering skill level
-    if (m_caster->GetTypeId() == TYPEID_PLAYER && m_CastItem)
-    {
-        ItemPrototype const* proto = m_CastItem->GetProto();
-        if (proto && proto->RequiredSkill == SKILL_ENGINEERING)
-            if (uint16 engineeringSkill = ((Player*)m_caster)->GetSkillValue(SKILL_ENGINEERING))
-            {
-                level = engineeringSkill / 5;
-                amount = 1;                                 // TODO HACK (needs a neat way of doing)
-            }
-    }
-
-    CreatureSummonPositions summonPositions;
-    summonPositions.resize(amount, CreaturePosition());
-
-    // Set middle position
-    if (m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION)
-    {
-        m_targets.getDestination(summonPositions[0].x, summonPositions[0].y, summonPositions[0].z);
-    }
-    else
-    {
-        realCaster->GetPosition(summonPositions[0].x, summonPositions[0].y, summonPositions[0].z);
-
-        // TODO - Is this really an error?
-        sLog.outDebug("Spell Effect EFFECT_SUMMON (%u) - summon without destination (spell id %u, effIndex %u)", effect->Effect, m_spellInfo->Id, SpellEffectIndex(effect->EffectIndex));
-    }
-
-    // Set summon positions
-    float radius = GetSpellRadius(sSpellRadiusStore.LookupEntry(effect->GetRadiusIndex()));
-    CreatureSummonPositions::iterator itr = summonPositions.begin();
-    for (++itr; itr != summonPositions.end(); ++itr)        // In case of multiple summons around position for not-fist positions
-    {
-        if (m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION || radius > 1.0f)
-        {
-            realCaster->GetRandomPoint(summonPositions[0].x, summonPositions[0].y, summonPositions[0].z, radius, itr->x, itr->y, itr->z);
-            if (realCaster->GetMap()->GetHitPosition(summonPositions[0].x, summonPositions[0].y, summonPositions[0].z, itr->x, itr->y, itr->z, m_caster->GetPhaseMask(), -0.5f))
-            {
-                realCaster->UpdateAllowedPositionZ(itr->x, itr->y, itr->z);
-            }
-        }
-        else                                                // Get a point near the caster
-        {
-            realCaster->GetRandomPoint(summonPositions[0].x, summonPositions[0].y, summonPositions[0].z, radius, itr->x, itr->y, itr->z);
-            if (realCaster->GetMap()->GetHitPosition(summonPositions[0].x, summonPositions[0].y, summonPositions[0].z, itr->x, itr->y, itr->z, m_caster->GetPhaseMask(), -0.5f))
-            {
-                realCaster->UpdateAllowedPositionZ(itr->x, itr->y, itr->z);
-            }
-        }
-    }
-
-    bool summonResult = false;
-    switch (summon_prop->Group)
-    {
-            // faction handled later on, or loaded from template
-        case SUMMON_PROP_GROUP_WILD:
-        case SUMMON_PROP_GROUP_FRIENDLY:
-        {
-            switch (summon_prop->Title)                     // better from known way sorting summons by AI types
-            {
-                case UNITNAME_SUMMON_TITLE_NONE:
-                {
-                    // those are classical totems - effectbasepoints is their hp and not summon ammount!
-                    // 121: 23035, battlestands
-                    // 647: 52893, Anti-Magic Zone (npc used)
-                    if (prop_id == 121 || prop_id == 647)
-                    {
-                        summonResult = DoSummonTotem(effect);
-                    }
-                    else
-                    {
-                        summonResult = DoSummonWild(summonPositions, summon_prop, effect, level);
-                    }
-                    break;
-                }
-                case UNITNAME_SUMMON_TITLE_PET:
-                case UNITNAME_SUMMON_TITLE_MINION:
-                case UNITNAME_SUMMON_TITLE_RUNEBLADE:
-                    summonResult = DoSummonGuardian(summonPositions, summon_prop, effect, level);
-                    break;
-                case UNITNAME_SUMMON_TITLE_GUARDIAN:
-                {
-                    if (prop_id == 61)                      // mixed guardians, totems, statues
-                    {
-                        // * Stone Statue, etc  -- fits much better totem AI
-                        if (m_spellInfo->SpellIconID == 2056)
-                        {
-                            summonResult = DoSummonTotem(effect);
-                        }
-                        else
-                        {
-                            // possible sort totems/guardians only by summon creature type
-                            CreatureInfo const* cInfo = ObjectMgr::GetCreatureTemplate(effect->EffectMiscValue);
-
-                            if (!cInfo)
-                            {
-                                return;
-                            }
-
-                            // FIXME: not all totems and similar cases selected by this check...
-                            if (cInfo->CreatureType == CREATURE_TYPE_TOTEM)
-                            {
-                                summonResult = DoSummonTotem(effect);
-                            }
-                            else
-                            {
-                                summonResult = DoSummonGuardian(summonPositions, summon_prop, effect, level);
-                            }
-                        }
-                    }
-                    else
-                    {
-                        summonResult = DoSummonGuardian(summonPositions, summon_prop, effect, level);
-                    }
-                    break;
-                }
-                case UNITNAME_SUMMON_TITLE_CONSTRUCT:
-                {
-                    if (prop_id == 2913)                    // Scrapbot
-                    {
-                        summonResult = DoSummonWild(summonPositions, summon_prop, effect, level);
-                    }
-                    else
-                    {
-                        summonResult = DoSummonGuardian(summonPositions, summon_prop, effect, level);
-                    }
-                    break;
-                }
-                case UNITNAME_SUMMON_TITLE_TOTEM:
-                    summonResult = DoSummonTotem(effect, summon_prop->Slot);
-                    break;
-                case UNITNAME_SUMMON_TITLE_COMPANION:
-                    // slot 6 set for critters that can help to player in fighting
-                    if (summon_prop->Slot == 6)
-                    {
-                        summonResult = DoSummonGuardian(summonPositions, summon_prop, effect, level);
-                    }
-                    else
-                    {
-                        summonResult = DoSummonCritter(summonPositions, summon_prop, effect, level);
-                    }
-                    break;
-                case UNITNAME_SUMMON_TITLE_OPPONENT:
-                case UNITNAME_SUMMON_TITLE_VEHICLE:
-                case UNITNAME_SUMMON_TITLE_MOUNT:
-                case UNITNAME_SUMMON_TITLE_LIGHTWELL:
-                case UNITNAME_SUMMON_TITLE_BUTLER:
-                    summonResult = DoSummonWild(summonPositions, summon_prop, effect, level);
-                    break;
-                default:
-                    sLog.outError("EffectSummonType: Unhandled summon title %u", summon_prop->Title);
-                    break;
-            }
-            break;
-        }
-        case SUMMON_PROP_GROUP_PETS:
-        {
-            // FIXME : multiple summons -  not yet supported as pet
-            // 1562 - force of nature  - sid 33831
-            // 1161 - feral spirit - sid 51533
-            if (prop_id == 1562)                            // 3 uncontrolable instead of one controllable :/
-            {
-                summonResult = DoSummonGuardian(summonPositions, summon_prop, effect, level);
-            }
-            break;
-        }
-        case SUMMON_PROP_GROUP_CONTROLLABLE:
-        {
-            summonResult = DoSummonPossessed(summonPositions, summon_prop, effect, level);
-            break;
-        }
-        case SUMMON_PROP_GROUP_VEHICLE:
-        case SUMMON_PROP_GROUP_UNCONTROLLABLE_VEHICLE:
-        {
-            summonResult = DoSummonVehicle(summonPositions, summon_prop, effect, level);
-            break;
-        }
-        default:
-            sLog.outError("EffectSummonType: Unhandled summon group type %u", summon_prop->Group);
-            break;
-    }
-
-    if (!summonResult)
-    {
-        return;                                             // No further handling required
-    }
-
-    for (itr = summonPositions.begin(); itr != summonPositions.end(); ++itr)
-    {
-        MANGOS_ASSERT(itr->creature || itr != summonPositions.begin());
-        if (!itr->creature)
-        {
-            sLog.outError("EffectSummonType: Expected to have %u NPCs summoned, but some failed (Spell id %u)", amount, m_spellInfo->Id);
-            continue;
-        }
-
-        if (summon_prop->FactionId)
-        {
-            itr->creature->setFaction(summon_prop->FactionId);
-        }
-        // Else set faction to summoner's faction for pet-like summoned
-        else if ((summon_prop->Flags & SUMMON_PROP_FLAG_INHERIT_FACTION) || !itr->creature->IsTemporarySummon())
-        {
-            itr->creature->setFaction(responsibleCaster ? responsibleCaster->getFaction() : m_caster->getFaction());
-        }
-
-        if (!itr->creature->IsTemporarySummon())
-        {
-            itr->creature->AIM_Initialize();
-
-            m_caster->GetMap()->Add(itr->creature);
-
-            // Notify Summoner
-            if (m_caster->GetTypeId() == TYPEID_UNIT && ((Creature*)m_caster)->AI())
-            {
-                ((Creature*)m_caster)->AI()->JustSummoned(itr->creature);
-            }
-            if (m_originalCaster && m_originalCaster != m_caster && m_originalCaster->GetTypeId() == TYPEID_UNIT && ((Creature*)m_originalCaster)->AI())
-            {
-                ((Creature*)m_originalCaster)->AI()->JustSummoned(itr->creature);
-            }
-
-            // used by eluna
-#ifdef ENABLE_ELUNA
-            if (Unit* summoner = m_caster->ToUnit())
-            {
-                if (Eluna* e = summoner->GetEluna())
-                {
-                    e->OnSummoned(itr->creature, summoner);
-                }
-            }
-            else if (m_originalCaster)
-            {
-                if (Unit* summoner = m_originalCaster->ToUnit())
-                {
-                    if (Eluna* e = summoner->GetEluna())
-                    {
-                        e->OnSummoned(itr->creature, summoner);
-                    }
-                }
-            }
-#endif
-        }
-    }
-}
-
-bool Spell::DoSummonWild(CreatureSummonPositions& list, SummonPropertiesEntry const* prop, SpellEffectEntry const* effect, uint32 level)
-{
-    MANGOS_ASSERT(!list.empty() && prop);
-
-    uint32 creature_entry = effect->EffectMiscValue;
-    CreatureInfo const* cInfo = ObjectMgr::GetCreatureTemplate(creature_entry);
-    if (!cInfo)
-    {
-        sLog.outErrorDb("Spell::DoSummonWild: creature entry %u not found for spell %u.", creature_entry, m_spellInfo->Id);
-        return false;
-    }
-
-    TempSpawnType summonType = (m_duration == 0) ? TEMPSPAWN_DEAD_DESPAWN : TEMPSPAWN_TIMED_OOC_OR_DEAD_DESPAWN;
-
-    for (CreatureSummonPositions::iterator itr = list.begin(); itr != list.end(); ++itr)
-        if (Creature* summon = m_caster->SummonCreature(creature_entry, itr->x, itr->y, itr->z, m_caster->GetOrientation(), summonType, m_duration))
-        {
-            itr->creature = summon;
-
-            summon->SetUInt32Value(UNIT_CREATED_BY_SPELL, m_spellInfo->Id);
-
-            // UNIT_FIELD_CREATEDBY are not set for these kind of spells.
-            // Does exceptions exist? If so, what are they?
-            // summon->SetCreatorGuid(m_caster->GetObjectGuid());
-
-            // Notify original caster if not done already
-            if (m_originalCaster && m_originalCaster != m_caster && m_originalCaster->GetTypeId() == TYPEID_UNIT && ((Creature*)m_originalCaster)->AI())
-            {
-                ((Creature*)m_originalCaster)->AI()->JustSummoned(summon);
-#ifdef ENABLE_ELUNA
-                if (Unit* summoner = m_originalCaster->ToUnit())
-                {
-                    if (Eluna* e = summoner->GetEluna())
-                    {
-                        e->OnSummoned(summon, summoner);
-                    }
-                }
-#endif
-            }
-        }
-        else
-        {
-            return false;
-        }
-
-    return true;
-}
-
-bool Spell::DoSummonCritter(CreatureSummonPositions& list, SummonPropertiesEntry const* prop, SpellEffectEntry const* effect, uint32 /*level*/)
-{
-    MANGOS_ASSERT(!list.empty() && prop);
-
-    // ATM only first position is supported for summoning
-    uint32 pet_entry = effect->EffectMiscValue;
-    CreatureInfo const* cInfo = ObjectMgr::GetCreatureTemplate(pet_entry);
-    if (!cInfo)
-    {
-        sLog.outErrorDb("Spell::DoSummonCritter: creature entry %u not found for spell %u.", pet_entry, m_spellInfo->Id);
-        return false;
-    }
-
-    Pet* old_critter = m_caster->GetMiniPet();
-
-    // for same pet just despawn (player unsummon command)
-    if (m_caster->GetTypeId() == TYPEID_PLAYER && old_critter && old_critter->GetEntry() == pet_entry)
-    {
-        m_caster->RemoveMiniPet();
-        return false;
-    }
-
-    // despawn old pet before summon new
-    if (old_critter)
-    {
-        m_caster->RemoveMiniPet();
-    }
-
-    // for (CreatureSummonPositions::iterator itr = list.begin(); itr != list.end(); ++itr)
-    CreatureCreatePos pos(m_caster->GetMap(), list[0].x, list[0].y, list[0].z, m_caster->GetOrientation(), m_caster->GetPhaseMask());
-
-    // summon new pet
-    Pet* critter = new Pet(MINI_PET);
-
-    uint32 pet_number = sObjectMgr.GeneratePetNumber();
-    if (!critter->Create(m_caster->GetMap()->GenerateLocalLowGuid(HIGHGUID_PET), pos, cInfo, pet_number))
-    {
-        sLog.outError("Spell::EffectSummonCritter, spellid %u: no such creature entry %u", m_spellInfo->Id, pet_entry);
-        delete critter;
-        return false;
-    }
-
-    // itr!
-    list[0].creature = critter;
-
-    critter->SetRespawnCoord(pos);
-
-    // critter->SetName("");                                // generated by client
-    critter->SetOwnerGuid(m_caster->GetObjectGuid());
-    critter->SetCreatorGuid(m_caster->GetObjectGuid());
-
-    critter->SetUInt32Value(UNIT_CREATED_BY_SPELL, m_spellInfo->Id);
-
-    critter->InitPetCreateSpells();                         // e.g. disgusting oozeling has a create spell as critter...
-    // critter->InitLevelupSpellsForLevel();                // none?
-    critter->SelectLevel(critter->GetCreatureInfo());       // some summoned creaters have different from 1 DB data for level/hp
-    critter->SetUInt32Value(UNIT_NPC_FLAGS, critter->GetCreatureInfo()->NpcFlags);
-    // some mini-pets have quests
-    // set timer for unsummon
-    if (m_duration > 0)
-    {
-        critter->SetDuration(m_duration);
-    }
-
-    m_caster->SetMiniPet(critter);
-
-#ifdef ENABLE_ELUNA
-    if (Unit* summoner = m_caster->ToUnit())
-    {
-        if (Eluna* e = summoner->GetEluna())
-        {
-            e->OnSummoned(critter, summoner);
-        }
-    }
-    if (m_originalCaster)
-        if (Unit* summoner = m_originalCaster->ToUnit())
-        {
-            if (Eluna* e = summoner->GetEluna())
-            {
-                e->OnSummoned(critter, summoner);
-            }
-        }
-#endif
-
-    return true;
-}
-
-bool Spell::DoSummonGuardian(CreatureSummonPositions& list, SummonPropertiesEntry const* prop, SpellEffectEntry const* effect, uint32 level)
-{
-    MANGOS_ASSERT(!list.empty() && prop);
-
-    uint32 pet_entry = effect->EffectMiscValue;
-    CreatureInfo const* cInfo = ObjectMgr::GetCreatureTemplate(pet_entry);
-    if (!cInfo)
-    {
-        sLog.outErrorDb("Spell::DoSummonGuardian: creature entry %u not found for spell %u.", pet_entry, m_spellInfo->Id);
-        return false;
-    }
-
-    PetType petType = prop->Title == UNITNAME_SUMMON_TITLE_COMPANION ? PROTECTOR_PET : GUARDIAN_PET;
-
-    // second direct cast unsummon guardian(s) (guardians without like functionality have cooldown > spawn time)
-    if (!m_IsTriggeredSpell && m_caster->GetTypeId() == TYPEID_PLAYER)
-    {
-        bool found = false;
-        // including protector
-        while (Pet* old_summon = m_caster->FindGuardianWithEntry(pet_entry))
-        {
-            old_summon->Unsummon(PET_SAVE_AS_DELETED, m_caster);
-            found = true;
-        }
-
-        if (found)
-        {
-            return false;
-        }
-    }
-
-    // protectors allowed only in single amount
-    if (petType == PROTECTOR_PET)
-        if (Pet* old_protector = m_caster->GetProtectorPet())
-        {
-            old_protector->Unsummon(PET_SAVE_AS_DELETED, m_caster);
-        }
-
-    // in another case summon new
-    for (CreatureSummonPositions::iterator itr = list.begin(); itr != list.end(); ++itr)
-    {
-        Pet* spawnCreature = new Pet(petType);
-
-        CreatureCreatePos pos(m_caster->GetMap(), itr->x, itr->y, itr->z, -m_caster->GetOrientation(), m_caster->GetPhaseMask());
-
-        uint32 pet_number = sObjectMgr.GeneratePetNumber();
-        if (!spawnCreature->Create(m_caster->GetMap()->GenerateLocalLowGuid(HIGHGUID_PET), pos, cInfo, pet_number))
-        {
-            sLog.outError("Spell::DoSummonGuardian: can't create creature entry %u for spell %u.", pet_entry, m_spellInfo->Id);
-            delete spawnCreature;
-            return false;
-        }
-
-        itr->creature = spawnCreature;
-
-        spawnCreature->SetRespawnCoord(pos);
-
-        if (m_duration > 0)
-        {
-            spawnCreature->SetDuration(m_duration);
-        }
-
-        CreatureInfo const* cInfo = spawnCreature->GetCreatureInfo();
-
-        // spawnCreature->SetName("");                      // generated by client
-        spawnCreature->SetOwnerGuid(m_caster->GetObjectGuid());
-        spawnCreature->SetUInt32Value(UNIT_FIELD_FLAGS, cInfo->UnitFlags);
-        spawnCreature->SetUInt32Value(UNIT_NPC_FLAGS, cInfo->NpcFlags);
-        spawnCreature->setFaction(m_caster->getFaction());
-        spawnCreature->SetUInt32Value(UNIT_FIELD_PET_NAME_TIMESTAMP, 0);
-        spawnCreature->SetCreatorGuid(m_caster->GetObjectGuid());
-        spawnCreature->SetUInt32Value(UNIT_CREATED_BY_SPELL, m_spellInfo->Id);
-
-        spawnCreature->InitStatsForLevel(level);
-
-        if (CharmInfo* charmInfo = spawnCreature->GetCharmInfo())
-        {
-            charmInfo->SetPetNumber(pet_number, false);
-
-            if (spawnCreature->HasFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_PASSIVE))
-            {
-                charmInfo->SetReactState(REACT_PASSIVE);
-            }
-            else if ((cInfo->ExtraFlags & CREATURE_FLAG_EXTRA_NO_MELEE) || petType == PROTECTOR_PET)
-            {
-                charmInfo->SetReactState(REACT_DEFENSIVE);
-            }
-            else
-            {
-                charmInfo->SetReactState(REACT_AGGRESSIVE);
-            }
-        }
-
-        m_caster->AddGuardian(spawnCreature);
-
-#ifdef ENABLE_ELUNA
-        if (Unit* summoner = m_caster->ToUnit())
-        {
-            if (Eluna* e = summoner->GetEluna())
-            {
-                e->OnSummoned(spawnCreature, summoner);
-            }
-        }
-        if (m_originalCaster)
-        {
-            if (Unit* summoner = m_originalCaster->ToUnit())
-            {
-                if (Eluna* e = summoner->GetEluna())
-                {
-                    e->OnSummoned(spawnCreature, summoner);
-                }
-            }
-        }
-#endif
-    }
-
-    return true;
-}
-
-bool Spell::DoSummonTotem(SpellEffectEntry const* effect, uint8 slot_dbc)
-{
-    // DBC store slots starting from 1, with no slot 0 value)
-    int slot = slot_dbc ? slot_dbc - 1 : TOTEM_SLOT_NONE;
-
-    // unsummon old totem
-    if (slot < MAX_TOTEM_SLOT)
-        if (Totem* OldTotem = m_caster->GetTotem(TotemSlot(slot)))
-        {
-            OldTotem->UnSummon();
-        }
-
-    // FIXME: Setup near to finish point because GetObjectBoundingRadius set in Create but some Create calls can be dependent from proper position
-    // if totem have creature_template_addon.auras with persistent point for example or script call
-    float angle = slot < MAX_TOTEM_SLOT ? M_PI_F / MAX_TOTEM_SLOT - (slot * 2 * M_PI_F / MAX_TOTEM_SLOT) : 0;
-
-    CreatureCreatePos pos(m_caster, m_caster->GetOrientation(), 2.0f, angle);
-
-    CreatureInfo const* cinfo = ObjectMgr::GetCreatureTemplate(effect->EffectMiscValue);
-    if (!cinfo)
-    {
-        sLog.outErrorDb("Creature entry %u does not exist but used in spell %u totem summon.", m_spellInfo->Id, effect->EffectMiscValue);
-        return false;
-    }
-
-    Totem* pTotem = new Totem;
-
-    if (!pTotem->Create(m_caster->GetMap()->GenerateLocalLowGuid(HIGHGUID_UNIT), pos, cinfo, m_caster))
-    {
-        delete pTotem;
-        return false;
-    }
-
-    pTotem->SetRespawnCoord(pos);
-
-    if (slot < MAX_TOTEM_SLOT)
-    {
-        m_caster->_AddTotem(TotemSlot(slot), pTotem);
-    }
-
-    // pTotem->SetName("");                                 // generated by client
-    pTotem->SetOwner(m_caster);
-    pTotem->SetTypeBySummonSpell(m_spellInfo);              // must be after Create call where m_spells initialized
-
-    pTotem->SetDuration(m_duration);
-
-    if (damage)                                             // if not spell info, DB values used
-    {
-        pTotem->SetMaxHealth(damage);
-        pTotem->SetHealth(damage);
-    }
-
-    pTotem->SetUInt32Value(UNIT_CREATED_BY_SPELL, m_spellInfo->Id);
-
-    if (m_caster->GetTypeId() == TYPEID_PLAYER)
-    {
-        pTotem->SetFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_PVP_ATTACKABLE);
-    }
-
-    if (m_caster->IsPvP())
-    {
-        pTotem->SetPvP(true);
-    }
-
-    if (m_caster->IsFFAPvP())
-    {
-        pTotem->SetFFAPvP(true);
-    }
-
-    // sending SMSG_TOTEM_CREATED before add to map (done in Summon)
-    if (slot < MAX_TOTEM_SLOT && m_caster->GetTypeId() == TYPEID_PLAYER)
-    {
-        WorldPacket data(SMSG_TOTEM_CREATED, 1 + 8 + 4 + 4);
-        data << uint8(slot);
-        data << pTotem->GetObjectGuid();
-        data << uint32(m_duration);
-        data << uint32(m_spellInfo->Id);
-        ((Player*)m_caster)->SendDirectMessage(&data);
-    }
-
-    pTotem->Summon(m_caster);
-
-    return false;
-}
-
-bool Spell::DoSummonPossessed(CreatureSummonPositions& list, SummonPropertiesEntry const* prop, SpellEffectEntry const* effect, uint32 level)
-{
-    MANGOS_ASSERT(!list.empty() && prop);
-
-    uint32 const& creatureEntry = effect->EffectMiscValue;
-
-    Unit* newUnit = m_caster->TakePossessOf(m_spellInfo, prop, effect, list[0].x, list[0].y, list[0].z, m_caster->GetOrientation());
-    if (!newUnit)
-    {
-        sLog.outError("Spell::DoSummonPossessed: creature entry %d for spell %u could not be summoned.", creatureEntry, m_spellInfo->Id);
-        return false;
-    }
-
-    list[0].creature = static_cast<Creature*>(newUnit);
-
-    // Notify Summoner
-    if (m_originalCaster && m_originalCaster != m_caster && m_originalCaster->GetTypeId() == TYPEID_UNIT && ((Creature*)m_originalCaster)->AI())
-    {
-        ((Creature*)m_originalCaster)->AI()->JustSummoned(list[0].creature);
-
-#ifdef ENABLE_ELUNA
-        if (Unit* summoner = m_originalCaster->ToUnit())
-        {
-            if (Eluna* e = summoner->GetEluna())
-            {
-                e->OnSummoned(list[0].creature, summoner);
-            }
-        }
-#endif
-    }
-
-    return true;
-}
-
-bool Spell::DoSummonPet(SpellEffectEntry const* effect)
-{
-    if (m_caster->GetPetGuid())
-    {
-        return false;
-    }
-
-    if (!unitTarget)
-    {
-        return false;
-    }
-
-    uint32 pet_entry = effect->EffectMiscValue;
-    CreatureInfo const* cInfo = ObjectMgr::GetCreatureTemplate(pet_entry);
-    if (!cInfo)
-    {
-        sLog.outErrorDb("Spell::DoSummonPet: creature entry %u not found for spell %u.", pet_entry, m_spellInfo->Id);
-        return false;
-    }
-
-    Pet* spawnCreature = new Pet();
-
-    // set timer for unsummon
-    if (m_duration > 0)
-    {
-        spawnCreature->SetDuration(m_duration);
-    }
-
-    if (m_caster->GetTypeId() == TYPEID_PLAYER)
-    {
-        if (spawnCreature->LoadPetFromDB((Player*)m_caster, pet_entry))
-        {
-            // Summon in dest location
-            if (m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION)
-            {
-                spawnCreature->Relocate(m_targets.m_destX, m_targets.m_destY, m_targets.m_destZ, -m_caster->GetOrientation());
-            }
-
-            return true;
-        }
-
-        spawnCreature->setPetType(SUMMON_PET);
-    }
-    else
-    {
-        spawnCreature->setPetType(GUARDIAN_PET);
-    }
-
-    // Summon in dest location
-    CreatureCreatePos pos(m_caster->GetMap(), m_targets.m_destX, m_targets.m_destY, m_targets.m_destZ, -m_caster->GetOrientation(), m_caster->GetPhaseMask());
-
-    if (!(m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION))
-    {
-        pos = CreatureCreatePos(m_caster, -m_caster->GetOrientation());
-    }
-
-    Map* map = m_caster->GetMap();
-    uint32 pet_number = sObjectMgr.GeneratePetNumber();
-    if (!spawnCreature->Create(map->GenerateLocalLowGuid(HIGHGUID_PET), pos, cInfo, pet_number))
-    {
-        sLog.outErrorDb("Spell::EffectSummon: can't create creature with entry %u for spell %u", cInfo->Entry, m_spellInfo->Id);
-        delete spawnCreature;
-        return false;
-    }
-
-    uint32 level = std::max(m_caster->getLevel() + effect->EffectMultipleValue, 1.0f);
-
-    spawnCreature->SetRespawnCoord(pos);
-
-    spawnCreature->SetOwnerGuid(m_caster->GetObjectGuid());
-    spawnCreature->setFaction(m_caster->getFaction());
-    spawnCreature->SetUInt32Value(UNIT_FIELD_PET_NAME_TIMESTAMP, 0);
-    spawnCreature->SetCreatorGuid(m_caster->GetObjectGuid());
-    spawnCreature->SetUInt32Value(UNIT_CREATED_BY_SPELL, m_spellInfo->Id);
-
-    spawnCreature->InitStatsForLevel(level);
-
-    spawnCreature->GetCharmInfo()->SetPetNumber(pet_number, false);
-
-    spawnCreature->InitPetCreateSpells();
-    spawnCreature->InitLevelupSpellsForLevel();
-
-    // spawnCreature->SetName("");                          // generated by client
-
-    map->Add((Creature*)spawnCreature);
-    spawnCreature->AIM_Initialize();
-
-    m_caster->SetPet(spawnCreature);
-
-    if (m_caster->GetTypeId() == TYPEID_PLAYER)
-    {
-        spawnCreature->GetCharmInfo()->SetReactState(REACT_DEFENSIVE);
-        spawnCreature->SavePetToDB(PET_SAVE_AS_CURRENT);
-        ((Player*)m_caster)->PetSpellInitialize();
-    }
-    else
-    {
-        // Notify Summoner
-        if (m_originalCaster && (m_originalCaster != m_caster)
-            && (m_originalCaster->GetTypeId() == TYPEID_UNIT) && ((Creature*)m_originalCaster)->AI())
-        {
-            ((Creature*)m_originalCaster)->AI()->JustSummoned(spawnCreature);
-            if (m_originalCaster->IsInCombat() && !(spawnCreature->HasFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_PASSIVE)))
-            {
-                ((Creature*)spawnCreature)->AI()->AttackStart(m_originalCaster->getAttackerForHelper());
-            }
-        }
-        else if ((m_caster->GetTypeId() == TYPEID_UNIT) && ((Creature*)m_caster)->AI())
-        {
-            ((Creature*)m_caster)->AI()->JustSummoned(spawnCreature);
-            if (m_caster->IsInCombat() && !(spawnCreature->HasFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_PASSIVE)))
-            {
-                ((Creature*)spawnCreature)->AI()->AttackStart(m_caster->getAttackerForHelper());
-            }
-        }
-    }
-
-#ifdef ENABLE_ELUNA
-    if (Unit* summoner = m_caster->ToUnit())
-    {
-        if (Eluna* e = summoner->GetEluna())
-        {
-            e->OnSummoned(spawnCreature, summoner);
-        }
-    }
-    if (m_originalCaster)
-    {
-        if (Unit* summoner = m_originalCaster->ToUnit())
-        {
-            if (Eluna* e = summoner->GetEluna())
-            {
-                e->OnSummoned(spawnCreature, summoner);
-            }
-        }
-    }
-#endif
-    return true;
-}
-
-bool Spell::DoSummonVehicle(CreatureSummonPositions& list, SummonPropertiesEntry const* prop, SpellEffectEntry const * effect, uint32 /*level*/)
-{
-    MANGOS_ASSERT(!list.empty() && prop);
-
-    uint32 creatureEntry = effect->EffectMiscValue;
-    CreatureInfo const* cInfo = ObjectMgr::GetCreatureTemplate(creatureEntry);
-    if (!cInfo)
-    {
-        sLog.outErrorDb("Spell::DoSummonVehicle: creature entry %u not found for spell %u.", creatureEntry, m_spellInfo->Id);
-        return false;
-    }
-
-    Creature* spawnCreature = m_caster->SummonCreature(creatureEntry, list[0].x, list[0].y, list[0].z, m_caster->GetOrientation(),
-                              (m_duration == 0) ? TEMPSPAWN_CORPSE_DESPAWN : TEMPSPAWN_TIMED_OOC_OR_CORPSE_DESPAWN, m_duration);
-
-    if (!spawnCreature)
-    {
-        sLog.outError("Spell::DoSummonVehicle: creature entry %u for spell %u could not be summoned.", creatureEntry, m_spellInfo->Id);
-        return false;
-    }
-
-    list[0].creature = spawnCreature;
-
-    // Changes to be sent
-    spawnCreature->SetCreatorGuid(m_caster->GetObjectGuid());
-    spawnCreature->SetUInt32Value(UNIT_CREATED_BY_SPELL, m_spellInfo->Id);
-    //spawnCreature->SetLevel(level); // Do we need to set level for vehicles?
-
-    // Board the caster right after summoning
-    SpellEntry const* controlSpellEntry = sSpellStore.LookupEntry(effect->CalculateSimpleValue());
-    if (controlSpellEntry && IsSpellHaveAura(controlSpellEntry, SPELL_AURA_CONTROL_VEHICLE))
-    {
-        m_caster->CastSpell(spawnCreature, controlSpellEntry, true);
-    }
-    else
-    {
-        m_caster->CastSpell(spawnCreature, SPELL_RIDE_VEHICLE_HARDCODED, true);
-    }
-
-    // If the boarding failed...
-    if (!spawnCreature->HasAuraType(SPELL_AURA_CONTROL_VEHICLE))
-    {
-        spawnCreature->ForcedDespawn();
-        return false;
-    }
-
-    // Notify Summoner
-    if (m_originalCaster && m_originalCaster != m_caster && m_originalCaster->GetTypeId() == TYPEID_UNIT && ((Creature*)m_originalCaster)->AI())
-    {
-        ((Creature*)m_originalCaster)->AI()->JustSummoned(spawnCreature);
-    }
-#ifdef ENABLE_ELUNA
-    if (Unit* summoner = m_caster->ToUnit())
-    {
-        if (Eluna* e = summoner->GetEluna())
-        {
-            e->OnSummoned(spawnCreature, summoner);
-        }
-    }
-    else if (m_originalCaster)
-        if (Unit* summoner = m_originalCaster->ToUnit())
-        {
-            if (Eluna* e = summoner->GetEluna())
-            {
-                e->OnSummoned(spawnCreature, summoner);
-            }
-        }
-#endif /* ENABLE_ELUNA */
-    return true;
-}
-
-/**
- * @brief Adds flat threat from the caster to the unit target.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectThreat(SpellEffectEntry const* /*effect*/)
-{
-    if (!unitTarget || !unitTarget->IsAlive() || !m_caster->IsAlive())
-    {
-        return;
-    }
-
-    if (!unitTarget->CanHaveThreatList())
-    {
-        return;
-    }
-
-    unitTarget->AddThreat(m_caster, float(damage), false, GetSpellSchoolMask(m_spellInfo), m_spellInfo);
-}
-
-/**
- * @brief Heals the target for an amount equal to the caster's maximum health.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectHealMaxHealth(SpellEffectEntry const* /*effect*/)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-    if (!unitTarget->IsAlive())
-    {
-        return;
-    }
-
-    uint32 heal = m_caster->GetMaxHealth();
-
-    m_healing += heal;
-}
-
-/**
- * @brief Interrupts interruptible non-melee spells on the unit target.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectInterruptCast(SpellEffectEntry const* /*effect*/)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-    if (!unitTarget->IsAlive())
-    {
-        return;
-    }
-
-    // TODO: not all spells that used this effect apply cooldown at school spells
-    // also exist case: apply cooldown to interrupted cast only and to all spells
-    for (uint32 i = CURRENT_FIRST_NON_MELEE_SPELL; i < CURRENT_MAX_SPELL; ++i)
-    {
-        if (Spell* spell = unitTarget->GetCurrentSpell(CurrentSpellTypes(i)))
-        {
-            SpellEntry const* curSpellInfo = spell->m_spellInfo;
-            // check if we can interrupt spell
-            if ((curSpellInfo->GetInterruptFlags() & SPELL_INTERRUPT_FLAG_INTERRUPT) && curSpellInfo->GetPreventionType() == SPELL_PREVENTION_TYPE_SILENCE )
-            {
-                unitTarget->ProhibitSpellSchool(GetSpellSchoolMask(curSpellInfo), GetSpellDuration(m_spellInfo));
-                unitTarget->InterruptSpell(CurrentSpellTypes(i), false);
-            }
-        }
-    }
-}
-
-/**
- * @brief Summons a wild game object at the destination or near the caster.
- *
- * @param effect The summon object effect index.
- */
-void Spell::EffectSummonObjectWild(SpellEffectEntry const* effect)
-{
-    uint32 gameobject_id = effect->EffectMiscValue;
-
-    GameObject* pGameObj = new GameObject;
-
-    WorldObject* target = focusObject;
-    if (!target)
-    {
-        target = m_caster;
-    }
-
-    float x, y, z;
-    if (m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION)
-    {
-        m_targets.getDestination(x, y, z);
-    }
-    else
-    {
-        m_caster->GetClosePoint(x, y, z, DEFAULT_WORLD_OBJECT_SIZE);
-    }
-
-    Map* map = target->GetMap();
-
-    if (!pGameObj->Create(map->GenerateLocalLowGuid(HIGHGUID_GAMEOBJECT), gameobject_id, map,
-                          m_caster->GetPhaseMask(), x, y, z, target->GetOrientation()))
-    {
-        delete pGameObj;
-        return;
-    }
-
-    pGameObj->SetRespawnTime(m_duration > 0 ? m_duration / IN_MILLISECONDS : 0);
-    pGameObj->SetSpellId(m_spellInfo->Id);
-
-    // Wild object not have owner and check clickable by players
-    map->Add(pGameObj);
-
-    // Store the GO to the caster
-    m_caster->AddWildGameObject(pGameObj);
-
-    if (pGameObj->GetGoType() == GAMEOBJECT_TYPE_FLAGDROP && m_caster->GetTypeId() == TYPEID_PLAYER)
-    {
-        Player* pl = (Player*)m_caster;
-        BattleGround* bg = ((Player*)m_caster)->GetBattleGround();
-
-        switch (pGameObj->GetMapId())
-        {
-            case 489:                                       // WS
-            {
-                if (bg && bg->GetTypeID() == BATTLEGROUND_WS && bg->GetStatus() == STATUS_IN_PROGRESS)
-                {
-                    Team team = pl->GetTeam() == ALLIANCE ? HORDE : ALLIANCE;
-
-                    ((BattleGroundWS*)bg)->SetDroppedFlagGuid(pGameObj->GetObjectGuid(), team);
-                }
-                break;
-            }
-            case 566:                                       // EY
-            {
-                if (bg && bg->GetTypeID() == BATTLEGROUND_EY && bg->GetStatus() == STATUS_IN_PROGRESS)
-                {
-                    ((BattleGroundEY*)bg)->SetDroppedFlagGuid(pGameObj->GetObjectGuid());
-                }
-                break;
-            }
-        }
-    }
-
-    pGameObj->SummonLinkedTrapIfAny();
-
-    if (m_caster->GetTypeId() == TYPEID_UNIT && ((Creature*)m_caster)->AI())
-    {
-        ((Creature*)m_caster)->AI()->JustSummoned(pGameObj);
-    }
-    if (m_originalCaster && m_originalCaster != m_caster && m_originalCaster->GetTypeId() == TYPEID_UNIT && ((Creature*)m_originalCaster)->AI())
-    {
-        ((Creature*)m_originalCaster)->AI()->JustSummoned(pGameObj);
-    }
-}
-
-/**
- * @brief Clears combat and threat state for the target.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectSanctuary(SpellEffectEntry const* /*effect*/)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-    // unitTarget->CombatStop();
-
-    unitTarget->CombatStop();
-    unitTarget->GetHostileRefManager().deleteReferences();  // stop all fighting
-
-    // Vanish allows to remove all threat and cast regular stealth so other spells can be used
-    if (m_spellInfo->IsFitToFamily(SPELLFAMILY_ROGUE, UI64LIT(0x0000000000000800)))
-    {
-        ((Player*)m_caster)->RemoveSpellsCausingAura(SPELL_AURA_MOD_ROOT);
-    }
-}
-
-/**
- * @brief Adds combo points to the caster for the unit target.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectAddComboPoints(SpellEffectEntry const* effect /*effect*/)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-
-    if (m_caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    if (damage <= 0)
-    {
-        return;
-    }
-
-    ((Player*)m_caster)->AddComboPoints(unitTarget, damage);
-}
-
-/**
- * @brief Creates a duel flag object and starts a duel request between two players.
- *
- * @param effect The effect index containing the duel flag game object id.
- */
-void Spell::EffectDuel(SpellEffectEntry const* effect)
-{
-    if (!m_caster || !unitTarget || m_caster->GetTypeId() != TYPEID_PLAYER || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Player* caster = (Player*)m_caster;
-    Player* target = (Player*)unitTarget;
-
-    // caster or target already have requested duel
-    if (caster->duel || target->duel || !target->GetSocial() || target->GetSocial()->HasIgnore(caster->GetObjectGuid()))
-    {
-        return;
-    }
-
-    // Players can only fight a duel with each other outside (=not inside dungeons and not in capital cities)
-    AreaTableEntry const* casterAreaEntry = GetAreaEntryByAreaID(caster->GetAreaId());
-    if (casterAreaEntry && !(casterAreaEntry->flags & AREA_FLAG_DUEL))
-    {
-        SendCastResult(SPELL_FAILED_NO_DUELING);            // Dueling isn't allowed here
-        return;
-    }
-
-    AreaTableEntry const* targetAreaEntry = GetAreaEntryByAreaID(target->GetAreaId());
-    if (targetAreaEntry && !(targetAreaEntry->flags & AREA_FLAG_DUEL))
-    {
-        SendCastResult(SPELL_FAILED_NO_DUELING);            // Dueling isn't allowed here
-        return;
-    }
-
-    // CREATE DUEL FLAG OBJECT
-    GameObject* pGameObj = new GameObject;
-
-    uint32 gameobject_id = effect->EffectMiscValue;
-
-    Map* map = m_caster->GetMap();
-    float x = (m_caster->GetPositionX() + unitTarget->GetPositionX()) * 0.5f;
-    float y = (m_caster->GetPositionY() + unitTarget->GetPositionY()) * 0.5f;
-    float z = m_caster->GetPositionZ();
-    m_caster->UpdateAllowedPositionZ(x, y, z);
-    if (!pGameObj->Create(map->GenerateLocalLowGuid(HIGHGUID_GAMEOBJECT), gameobject_id, map, m_caster->GetPhaseMask(), x, y, z, m_caster->GetOrientation()))
-    {
-        delete pGameObj;
-        return;
-    }
-
-    pGameObj->SetUInt32Value(GAMEOBJECT_FACTION, m_caster->getFaction());
-    pGameObj->SetUInt32Value(GAMEOBJECT_LEVEL, m_caster->getLevel() + 1);
-
-    pGameObj->SetRespawnTime(m_duration > 0 ? m_duration / IN_MILLISECONDS : 0);
-    pGameObj->SetSpellId(m_spellInfo->Id);
-
-    m_caster->AddGameObject(pGameObj);
-    map->Add(pGameObj);
-    // END
-
-    // Send request
-    WorldPacket data(SMSG_DUEL_REQUESTED, 8 + 8);
-    data << pGameObj->GetObjectGuid();
-    data << caster->GetObjectGuid();
-    caster->GetSession()->SendPacket(&data);
-    target->GetSession()->SendPacket(&data);
-
-    // create duel-info
-    DuelInfo* duel   = new DuelInfo;
-    duel->initiator  = caster;
-    duel->opponent   = target;
-    duel->startTime  = 0;
-    duel->startTimer = 0;
-    caster->duel     = duel;
-
-    DuelInfo* duel2   = new DuelInfo;
-    duel2->initiator  = caster;
-    duel2->opponent   = caster;
-    duel2->startTime  = 0;
-    duel2->startTimer = 0;
-    target->duel      = duel2;
-
-    caster->SetGuidValue(PLAYER_DUEL_ARBITER, pGameObj->GetObjectGuid());
-    target->SetGuidValue(PLAYER_DUEL_ARBITER, pGameObj->GetObjectGuid());
-
-    // Used by Eluna
-#ifdef ENABLE_ELUNA
-    if (Eluna* e = caster->GetEluna())
-    {
-        e->OnDuelRequest(target, caster);
-    }
-#endif /* ENABLE_ELUNA */
-}
-
-/**
- * @brief Teleports a player target to its homebind as an unstuck action.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectStuck(SpellEffectEntry const* effect /*effect*/)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    if (!sWorld.getConfig(CONFIG_BOOL_CAST_UNSTUCK))
-    {
-        return;
-    }
-
-    Player* pTarget = (Player*)unitTarget;
-
-    DEBUG_LOG("Spell Effect: Stuck");
-    DETAIL_LOG("Player %s (guid %u) used auto-unstuck future at map %u (%f, %f, %f)", pTarget->GetName(), pTarget->GetGUIDLow(), m_caster->GetMapId(), m_caster->GetPositionX(), pTarget->GetPositionY(), pTarget->GetPositionZ());
-
-    if (pTarget->IsTaxiFlying())
-    {
-        return;
-    }
-
-    // homebind location is loaded always
-    pTarget->TeleportToHomebind(unitTarget == m_caster ? TELE_TO_SPELL : 0);
-
-    // Stuck spell trigger Hearthstone cooldown
-    SpellEntry const* spellInfo = sSpellStore.LookupEntry(8690);
-    if (!spellInfo)
-    {
-        return;
-    }
-    Spell spell(pTarget, spellInfo, true);
-    spell.SendSpellCooldown();
-}
-
-/**
- * @brief Sends a summon request to a player target.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectSummonPlayer(SpellEffectEntry const* /*effect*/)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    // Evil Twin (ignore player summon, but hide this for summoner)
-    if (unitTarget->GetDummyAura(23445))
-    {
-        return;
-    }
-
-    float x, y, z;
-    m_caster->GetClosePoint(x, y, z, unitTarget->GetObjectBoundingRadius());
-
-    ((Player*)unitTarget)->SetSummonPoint(m_caster->GetMapId(), x, y, z);
-
-    WorldPacket data(SMSG_SUMMON_REQUEST, 8 + 4 + 4);
-    data << m_caster->GetObjectGuid();                      // summoner guid
-    data << uint32(m_caster->GetZoneId());                  // summoner zone
-    data << uint32(MAX_PLAYER_SUMMON_DELAY * IN_MILLISECONDS); // auto decline after msecs
-    ((Player*)unitTarget)->GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Builds the default script command used to activate a game object.
- *
- * @return ScriptInfo Preconfigured activation command data.
- */
-static ScriptInfo generateActivateCommand()
-{
-    ScriptInfo si;
-    si.command = SCRIPT_COMMAND_ACTIVATE_OBJECT;
-    si.id = 0;
-    si.buddyEntry = 0;
-    si.searchRadiusOrGuid = 0;
-    si.data_flags = 0x00;
-    return si;
-}
-
-/**
- * @brief Activates or manipulates the targeted game object based on the effect misc value.
- *
- * @param effect The activation effect index.
- */
-void Spell::EffectActivateObject(SpellEffectEntry const* effect)
-{
-    if (!gameObjTarget)
-    {
-        return;
-    }
-
-    uint32 misc_value = uint32(effect->EffectMiscValue);
-
-    switch (misc_value)
-    {
-        case 1:                     // GO simple use
-        case 2:                     // unk - 2 spells
-        case 4:                     // unk - 1 spell
-        case 5:                     // GO trap usage
-        case 7:                     // unk - 2 spells
-        case 8:                     // GO usage with TargetB = none or random
-        case 10:                    // GO explosions
-        case 11:                    // unk - 1 spell
-        case 19:                    // unk - 1 spell
-        case 20:                    // unk - 2 spells
-        {
-            static ScriptInfo activateCommand = generateActivateCommand();
-
-            int32 delay_secs = effect->CalculateSimpleValue();
-            gameObjTarget->GetMap()->ScriptCommandStart(activateCommand, delay_secs, m_caster, gameObjTarget);
-            break;
-        }
-        case 3:                     // GO custom anim - found mostly in Lunar Fireworks spells
-            gameObjTarget->SendGameObjectCustomAnim(gameObjTarget->GetObjectGuid());
-            break;
-        case 12:                    // GO state active alternative - found mostly in Simon Game spells
-            gameObjTarget->UseDoorOrButton(0, true);
-            break;
-        case 13:                    // GO state ready - found only in Simon Game spells
-            gameObjTarget->ResetDoorOrButton();
-            break;
-        case 15:                    // GO destroy
-            gameObjTarget->SetLootState(GO_JUST_DEACTIVATED);
-            break;
-        case 16:                    // GO custom use - found mostly in Wind Stones spells, Simon Game spells and other GO target summoning spells
-        {
-            switch (m_spellInfo->Id)
-            {
-                case 24734:         // Summon Templar Random
-                case 24744:         // Summon Templar (fire)
-                case 24756:         // Summon Templar (air)
-                case 24758:         // Summon Templar (earth)
-                case 24760:         // Summon Templar (water)
-                case 24763:         // Summon Duke Random
-                case 24765:         // Summon Duke (fire)
-                case 24768:         // Summon Duke (air)
-                case 24770:         // Summon Duke (earth)
-                case 24772:         // Summon Duke (water)
-                case 24784:         // Summon Royal Random
-                case 24786:         // Summon Royal (fire)
-                case 24788:         // Summon Royal (air)
-                case 24789:         // Summon Royal (earth)
-                case 24790:         // Summon Royal (water)
-                {
-                    uint32 npcEntry = 0;
-                    uint32 templars[] = {15209, 15211, 15212, 15307};
-                    uint32 dukes[] = {15206, 15207, 15208, 15220};
-                    uint32 royals[] = {15203, 15204, 15205, 15305};
-
-                    switch (m_spellInfo->Id)
-                    {
-                        case 24734: npcEntry = templars[urand(0, 3)]; break;
-                        case 24763: npcEntry = dukes[urand(0, 3)];    break;
-                        case 24784: npcEntry = royals[urand(0, 3)];   break;
-                        case 24744: npcEntry = 15209;                 break;
-                        case 24756: npcEntry = 15212;                 break;
-                        case 24758: npcEntry = 15307;                 break;
-                        case 24760: npcEntry = 15211;                 break;
-                        case 24765: npcEntry = 15206;                 break;
-                        case 24768: npcEntry = 15220;                 break;
-                        case 24770: npcEntry = 15208;                 break;
-                        case 24772: npcEntry = 15207;                 break;
-                        case 24786: npcEntry = 15203;                 break;
-                        case 24788: npcEntry = 15204;                 break;
-                        case 24789: npcEntry = 15205;                 break;
-                        case 24790: npcEntry = 15305;                 break;
-                    }
-
-                    gameObjTarget->SummonCreature(npcEntry, gameObjTarget->GetPositionX(), gameObjTarget->GetPositionY(), gameObjTarget->GetPositionZ(), gameObjTarget->GetAngle(m_caster), TEMPSPAWN_TIMED_OOC_OR_DEAD_DESPAWN, MINUTE * IN_MILLISECONDS);
-                    gameObjTarget->SetLootState(GO_JUST_DEACTIVATED);
-                    break;
-                }
-                case 40176:         // Simon Game pre-game Begin, blue
-                case 40177:         // Simon Game pre-game Begin, green
-                case 40178:         // Simon Game pre-game Begin, red
-                case 40179:         // Simon Game pre-game Begin, yellow
-                case 40283:         // Simon Game END, blue
-                case 40284:         // Simon Game END, green
-                case 40285:         // Simon Game END, red
-                case 40286:         // Simon Game END, yellow
-                case 40494:         // Simon Game, switched ON
-                case 40495:         // Simon Game, switched OFF
-                case 40512:         // Simon Game, switch...disable Off switch
-                    gameObjTarget->SetFlag(GAMEOBJECT_FLAGS, GO_FLAG_NO_INTERACT);
-                    break;
-                case 40632:         // Summon Gezzarak the Huntress
-                case 40640:         // Summon Karrog
-                case 40642:         // Summon Darkscreecher Akkarai
-                case 40644:         // Summon Vakkiz the Windrager
-                case 41004:         // Summon Terokk
-                    gameObjTarget->SetLootState(GO_JUST_DEACTIVATED);
-                    break;
-                case 46085:         // Place Fake Fur
-                {
-                    float x, y, z;
-                    gameObjTarget->GetClosePoint(x, y, z, gameObjTarget->GetObjectBoundingRadius(), 2 * INTERACTION_DISTANCE, frand(0, M_PI_F * 2));
-
-                    // Note: event script is implemented in script library
-                    gameObjTarget->SummonCreature(25835, x, y, z, gameObjTarget->GetOrientation(), TEMPSPAWN_TIMED_OOC_OR_DEAD_DESPAWN, 15000);
-                    gameObjTarget->SetFlag(GAMEOBJECT_FLAGS, GO_FLAG_IN_USE);
-                    break;
-                }
-                case 46592:         // Summon Ahune Lieutenant
-                {
-                    uint32 npcEntry = 0;
-
-                    switch (gameObjTarget->GetEntry())
-                    {
-                        case 188049: npcEntry = 26116; break;       // Frostwave Lieutenant (Ashenvale)
-                        case 188137: npcEntry = 26178; break;       // Hailstone Lieutenant (Desolace)
-                        case 188138: npcEntry = 26204; break;       // Chillwind Lieutenant (Stranglethorn)
-                        case 188148: npcEntry = 26214; break;       // Frigid Lieutenant (Searing Gorge)
-                        case 188149: npcEntry = 26215; break;       // Glacial Lieutenant (Silithus)
-                        case 188150: npcEntry = 26216; break;       // Glacial Templar (Hellfire Peninsula)
-                    }
-
-                    gameObjTarget->SummonCreature(npcEntry, gameObjTarget->GetPositionX(), gameObjTarget->GetPositionY(), gameObjTarget->GetPositionZ(), gameObjTarget->GetAngle(m_caster), TEMPSPAWN_TIMED_OOC_OR_DEAD_DESPAWN, MINUTE * IN_MILLISECONDS);
-                    gameObjTarget->SetLootState(GO_JUST_DEACTIVATED);
-                    break;
-                }
-            }
-            break;
-        }
-        case 17:                    // GO unlock - found mostly in Simon Game spells
-            gameObjTarget->RemoveFlag(GAMEOBJECT_FLAGS, GO_FLAG_NO_INTERACT);
-            break;
-        default:
-            sLog.outError("Spell::EffectActivateObject called with unknown misc value. Spell Id %u", m_spellInfo->Id);
-            break;
-    }
-}
-
-void Spell::EffectApplyGlyph(SpellEffectEntry const* effect)
-{
-    if (m_caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Player* player = (Player*)m_caster;
-
-    // glyph sockets level requirement
-    uint8 minLevel = 0;
-    switch (m_glyphIndex)
-    {
-        case 0:
-        case 1:
-        case 6: minLevel = 25; break;
-        case 2:
-        case 3:
-        case 7: minLevel = 50; break;
-        case 4:
-        case 5:
-        case 8: minLevel = 75; break;
-    }
-
-    if (minLevel && m_caster->getLevel() < minLevel)
-    {
-        SendCastResult(SPELL_FAILED_GLYPH_SOCKET_LOCKED);
-        return;
-    }
-
-    // apply new one
-    if (uint32 glyph = effect->EffectMiscValue)
-    {
-        if (GlyphPropertiesEntry const* gp = sGlyphPropertiesStore.LookupEntry(glyph))
-        {
-            if (GlyphSlotEntry const* gs = sGlyphSlotStore.LookupEntry(player->GetGlyphSlot(m_glyphIndex)))
-            {
-                if (gp->TypeFlags != gs->TypeFlags)
-                {
-                    SendCastResult(SPELL_FAILED_INVALID_GLYPH);
-                    return;                                 // glyph slot mismatch
-                }
-            }
-
-            // remove old glyph
-            player->ApplyGlyph(m_glyphIndex, false);
-            player->SetGlyph(m_glyphIndex, glyph);
-            player->ApplyGlyph(m_glyphIndex, true);
-            player->SendTalentsInfoData(false);
-        }
-    }
-// TODO: ELUNAFIX NEEDED
-//#ifdef ENABLE_ELUNA
-//    if (Unit* summoner = m_originalCaster->ToUnit())
-//    {
-//        if (Eluna* e = player->GetEluna())
-//        {
-//            e->OnSummoned(spawnCreature, summoner);
-//        }
-//    }
-//#endif /* ENABLE_ELUNA */
-
-}
-
-/**
- * @brief Applies a temporary enchantment to the main-hand item of the player target.
- *
- * @param effect The enchant effect index.
- */
-void Spell::EffectEnchantHeldItem(SpellEffectEntry const* effect)
-{
-    // this is only item spell effect applied to main-hand weapon of target player (players in area)
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Player* item_owner = (Player*)unitTarget;
-    Item* item = item_owner->GetItemByPos(INVENTORY_SLOT_BAG_0, EQUIPMENT_SLOT_MAINHAND);
-
-    if (!item)
-    {
-        return;
-    }
-
-    // must be equipped
-    if (!item ->IsEquipped())
-    {
-        return;
-    }
-
-    if (effect->EffectMiscValue)
-    {
-        uint32 enchant_id = effect->EffectMiscValue;
-        int32 duration = m_duration;                        // Try duration index first...
-        if (!duration)
-        {
-            duration = m_currentBasePoints[SpellEffectIndex(effect->EffectIndex)];        // Base points after...
-        }
-        if (!duration)
-            duration = 10;                                  // 10 seconds for enchants which don't have listed duration
-
-        SpellItemEnchantmentEntry const* pEnchant = sSpellItemEnchantmentStore.LookupEntry(enchant_id);
-        if (!pEnchant)
-        {
-            return;
-        }
-
-        // Always go to temp enchantment slot
-        EnchantmentSlot slot = TEMP_ENCHANTMENT_SLOT;
-
-        // Enchantment will not be applied if a different one already exists
-        if (item->GetEnchantmentId(slot) && item->GetEnchantmentId(slot) != enchant_id)
-        {
-            return;
-        }
-
-        // Apply the temporary enchantment
-        item->SetEnchantment(slot, enchant_id, duration * IN_MILLISECONDS, 0, m_caster->GetObjectGuid());
-        item_owner->ApplyEnchantment(item, slot, true);
-    }
-}
-
-/**
- * @brief Starts disenchanting loot generation for the target item.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectDisEnchant(SpellEffectEntry const* /*effect*/)
-{
-    if (m_caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Player* p_caster = (Player*)m_caster;
-    if (!itemTarget || !itemTarget->GetProto()->DisenchantID)
-    {
-        return;
-    }
-
-    p_caster->UpdateCraftSkill(m_spellInfo->Id);
-
-    ((Player*)m_caster)->SendLoot(itemTarget->GetObjectGuid(), LOOT_DISENCHANTING);
-
-    // item will be removed at disenchanting end
-}
-
-/**
- * @brief Increases the drunk state of a player target.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectInebriate(SpellEffectEntry const* /*effect*/)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Player* player = (Player*)unitTarget;
-
-    uint8 drunkValue = player->GetDrunkValue() + (uint8)damage;
-    if (drunkValue > 100)
-    {
-        drunkValue = 100;
-        if (roll_chance_i(25))
-        {
-            player->CastSpell(player, 67468, false);    // Drunken Vomit
-        }
-    }
-    player->SetDrunkValue(drunkValue, m_CastItem ? m_CastItem->GetEntry() : 0);
-}
-
-/**
- * @brief Feeds the caster's pet and triggers the associated benefit spell.
- *
- * @param effect The effect index containing the triggered spell id.
- */
-void Spell::EffectFeedPet(SpellEffectEntry const* effect)
-{
-    if (m_caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Player* _player = (Player*)m_caster;
-
-    Item* foodItem = m_targets.getItemTarget();
-    if (!foodItem)
-    {
-        return;
-    }
-
-    Pet* pet = _player->GetPet();
-    if (!pet)
-    {
-        return;
-    }
-
-    if (!pet->IsAlive())
-    {
-        return;
-    }
-
-    int32 benefit = pet->GetCurrentFoodBenefitLevel(foodItem->GetProto()->ItemLevel);
-    if (benefit <= 0)
-    {
-        return;
-    }
-
-    uint32 count = 1;
-    _player->DestroyItemCount(foodItem, count, true);
-    // Clear item target to prevent dangling pointer if another effect targets the same item
-    m_targets.setItemTarget(NULL);
-
-    m_caster->CastCustomSpell(pet, effect->EffectTriggerSpell, &benefit, NULL, NULL, true);
-}
-
-/**
- * @brief Dismisses the caster's living pet.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectDismissPet(SpellEffectEntry const* /*effect*/)
-{
-    if (m_caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Pet* pet = m_caster->GetPet();
-
-    // not let dismiss dead pet
-    if (!pet || !pet->IsAlive())
-    {
-        return;
-    }
-
-    pet->Unsummon(PET_SAVE_NOT_IN_SLOT, m_caster);
-}
-
-/**
- * @brief Summons a persistent object into one of the caster's object slots.
- *
- * @param effect The summon object effect index.
- */
-void Spell::EffectSummonObject(SpellEffectEntry const* effect)
-{
-    uint32 go_id = effect->EffectMiscValue;
-    uint8 slot = effect->EffectMiscValueB;
-    if (slot >= MAX_OBJECT_SLOT)
-    {
-        return;
-    }
-
-    if (ObjectGuid guid = m_caster->m_ObjectSlotGuid[slot])
-    {
-        if (GameObject* obj = m_caster ? m_caster->GetMap()->GetGameObject(guid) : NULL)
-        {
-            obj->SetLootState(GO_JUST_DEACTIVATED);
-        }
-        m_caster->m_ObjectSlotGuid[slot].Clear();
-    }
-
-    GameObject* pGameObj = new GameObject;
-
-    float x, y, z;
-    // If dest location if present
-    if (m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION)
-    {
-        m_targets.getDestination(x, y, z);
-    }
-    // Summon in random point all other units if location present
-    else
-    {
-        m_caster->GetClosePoint(x, y, z, DEFAULT_WORLD_OBJECT_SIZE);
-    }
-
-    Map* map = m_caster->GetMap();
-    if (!pGameObj->Create(map->GenerateLocalLowGuid(HIGHGUID_GAMEOBJECT), go_id, map,
-                          m_caster->GetPhaseMask(), x, y, z, m_caster->GetOrientation()))
-    {
-        delete pGameObj;
-        return;
-    }
-
-    pGameObj->SetUInt32Value(GAMEOBJECT_LEVEL, m_caster->getLevel());
-    pGameObj->SetRespawnTime(m_duration > 0 ? m_duration / IN_MILLISECONDS : 0);
-    pGameObj->SetSpellId(m_spellInfo->Id);
-    m_caster->AddGameObject(pGameObj);
-
-    map->Add(pGameObj);
-
-    m_caster->m_ObjectSlotGuid[slot] = pGameObj->GetObjectGuid();
-
-    pGameObj->SummonLinkedTrapIfAny();
-
-    if (m_caster->GetTypeId() == TYPEID_UNIT && ((Creature*)m_caster)->AI())
-    {
-        ((Creature*)m_caster)->AI()->JustSummoned(pGameObj);
-    }
-    if (m_originalCaster && m_originalCaster != m_caster && m_originalCaster->GetTypeId() == TYPEID_UNIT && ((Creature*)m_originalCaster)->AI())
-    {
-        ((Creature*)m_originalCaster)->AI()->JustSummoned(pGameObj);
-    }
-}
-
-/**
- * @brief Sends a resurrection request with percentage-based health and mana restoration.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectResurrect(SpellEffectEntry const* /*effect*/)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    if (unitTarget->IsAlive() || !unitTarget->IsInWorld())
-    {
-        return;
-    }
-
-    switch (m_spellInfo->Id)
-    {
-        case 8342:                                          // Defibrillate (Goblin Jumper Cables) has 33% chance on success
-        case 22999:                                         // Defibrillate (Goblin Jumper Cables XL) has 50% chance on success
-        case 54732:                                         // Defibrillate (Gnomish Army Knife) has 67% chance on success
-        {
-            uint32 failChance = 0;
-            uint32 failSpellId = 0;
-            switch (m_spellInfo->Id)
-            {
-                case 8342:  failChance = 67; failSpellId = 8338;  break;
-                case 22999: failChance = 50; failSpellId = 23055; break;
-                case 54732: failChance = 33; failSpellId = 0; break;
-            }
-
-            if (roll_chance_i(failChance))
-            {
-                if (failSpellId)
-                {
-                    m_caster->CastSpell(m_caster, failSpellId, true, m_CastItem);
-                }
-                return;
-            }
-            break;
-        }
-        default:
-            break;
-    }
-
-    Player* pTarget = ((Player*)unitTarget);
-
-    if (pTarget->isRessurectRequested())      // already have one active request
-    {
-        return;
-    }
-
-    uint32 health = pTarget->GetMaxHealth() * damage / 100;
-    uint32 mana   = pTarget->GetMaxPower(POWER_MANA) * damage / 100;
-
-    pTarget->setResurrectRequestData(m_caster->GetObjectGuid(), m_caster->GetMapId(), m_caster->GetPositionX(), m_caster->GetPositionY(), m_caster->GetPositionZ(), health, mana);
-    SendResurrectRequest(pTarget);
-}
-
-/**
- * @brief Adds queued extra attacks to the target unit.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectAddExtraAttacks(SpellEffectEntry const* /*effect*/)
-{
-    if (!unitTarget || !unitTarget->IsAlive())
-    {
-        return;
-    }
-
-    if (unitTarget->m_extraAttacks)
-    {
-        return;
-    }
-
-    unitTarget->m_extraAttacks = damage;
-}
-
-/**
- * @brief Grants the ability to parry to a player target.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectParry(SpellEffectEntry const* /*effect*/)
-{
-    if (unitTarget && unitTarget->GetTypeId() == TYPEID_PLAYER)
-    {
-        ((Player*)unitTarget)->SetCanParry(true);
-    }
-}
-
-/**
- * @brief Grants the ability to block to a player target.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectBlock(SpellEffectEntry const* /*effect*/)
-{
-    if (unitTarget && unitTarget->GetTypeId() == TYPEID_PLAYER)
-    {
-        ((Player*)unitTarget)->SetCanBlock(true);
-    }
-}
-
-/**
- * @brief Teleports the target forward while avoiding steep terrain, water edges, and obstacles.
- *
- * @param effect The effect index providing the leap distance.
- */
-void Spell::EffectLeapForward(SpellEffectEntry const* effect)
-{
-    float dist = GetSpellRadius(sSpellRadiusStore.LookupEntry(effect->GetRadiusIndex()));
-    const float IN_OR_UNDER_LIQUID_RANGE = 0.8f;                // range to make player under liquid or on liquid surface from liquid level
-
-    G3D::Vector3 prevPos, nextPos;
-    float orientation = unitTarget->GetOrientation();
-
-    prevPos.x = unitTarget->GetPositionX();
-    prevPos.y = unitTarget->GetPositionY();
-    prevPos.z = unitTarget->GetPositionZ();
-
-    float groundZ = prevPos.z;
-
-    // falling case
-    if (!unitTarget->GetMap()->GetHeightInRange(unitTarget->GetPhaseMask(), prevPos.x, prevPos.y, groundZ, 3.0f) && unitTarget->m_movementInfo.HasMovementFlag(MOVEFLAG_FALLING))
-    {
-        nextPos.x = prevPos.x + dist * cos(orientation);
-        nextPos.y = prevPos.y + dist * sin(orientation);
-        nextPos.z = prevPos.z - 2.0f; // little hack to avoid the impression to go up when teleporting instead of continue to fall. This value may need some tweak
-
-        //
-        GridMapLiquidData liquidData;
-        if (unitTarget->GetMap()->GetTerrain()->IsInWater(nextPos.x, nextPos.y, nextPos.z, &liquidData))
-        {
-            if (fabs(nextPos.z - liquidData.level) < 10.0f)
-            {
-                nextPos.z = liquidData.level - IN_OR_UNDER_LIQUID_RANGE;
-            }
-        }
-        else
-        {
-            // fix z to ground if near of it
-            unitTarget->GetMap()->GetHeightInRange(unitTarget->GetPhaseMask(), nextPos.x, nextPos.y, nextPos.z, 10.0f);
-        }
-
-        // check any obstacle and fix coords
-        unitTarget->GetMap()->GetHitPosition(prevPos.x, prevPos.y, prevPos.z + 0.5f, nextPos.x, nextPos.y, nextPos.z, unitTarget->GetPhaseMask(), -0.5f);
-
-        // teleport
-        unitTarget->NearTeleportTo(nextPos.x, nextPos.y, nextPos.z, orientation, unitTarget == m_caster);
-
-        //sLog.outString("Falling BLINK!");
-        return;
-    }
-
-    // fix origin position if player was jumping and near of the ground but not in ground
-    if (fabs(prevPos.z - groundZ) > 0.5f)
-    {
-        prevPos.z = groundZ;
-    }
-
-    //check if in liquid
-    bool isPrevInLiquid = unitTarget->GetMap()->GetTerrain()->IsInWater(prevPos.x, prevPos.y, prevPos.z);
-
-    const float step = 2.0f;                                    // step length before next check slope/edge/water
-    const float maxSlope = 50.0f;                               // 50(degree) max seem best value for walkable slope
-    const float MAX_SLOPE_IN_RADIAN = maxSlope / 180.0f * M_PI_F;
-    float nextZPointEstimation = 1.0f;
-    float destx = prevPos.x + dist * cos(orientation);
-    float desty = prevPos.y + dist * sin(orientation);
-    const uint32 numChecks = ceil(fabs(dist / step));
-    const float DELTA_X = (destx - prevPos.x) / numChecks;
-    const float DELTA_Y = (desty - prevPos.y) / numChecks;
-
-    for (uint32 i = 1; i < numChecks + 1; ++i)
-    {
-        // compute next point average position
-        nextPos.x = prevPos.x + DELTA_X;
-        nextPos.y = prevPos.y + DELTA_Y;
-        nextPos.z = prevPos.z + nextZPointEstimation;
-
-        bool isInLiquid = false;
-        bool isInLiquidTested = false;
-        bool isOnGround = false;
-        GridMapLiquidData liquidData;
-
-        // try fix height for next position
-        if (!unitTarget->GetMap()->GetHeightInRange(unitTarget->GetPhaseMask(), nextPos.x, nextPos.y, nextPos.z))
-        {
-            // we cant so test if we are on water
-            if (!unitTarget->GetMap()->GetTerrain()->IsInWater(nextPos.x, nextPos.y, nextPos.z, &liquidData))
-            {
-                // not in water and cannot get correct height, maybe flying?
-                //sLog.outString("Can't get height of point %u, point value %s", i, nextPos.toString().c_str());
-                nextPos = prevPos;
-                break;
-            }
-            else
-            {
-                isInLiquid = true;
-                isInLiquidTested = true;
-            }
-        }
-        else
-        {
-            isOnGround = true;                                  // player is on ground
-        }
-
-        if (isInLiquid || (!isInLiquidTested && unitTarget->GetMap()->GetTerrain()->IsInWater(nextPos.x, nextPos.y, nextPos.z, &liquidData)))
-        {
-            if (!isPrevInLiquid && fabs(liquidData.level - prevPos.z) > 2.0f)
-            {
-                // on edge of water with difference a bit to high to continue
-                //sLog.outString("Ground vs liquid edge detected!");
-                nextPos = prevPos;
-                break;
-            }
-
-            if ((liquidData.level - IN_OR_UNDER_LIQUID_RANGE) > nextPos.z)
-            {
-                nextPos.z = prevPos.z;                                      // we are under water so next z equal prev z
-            }
-            else
-            {
-                nextPos.z = liquidData.level - IN_OR_UNDER_LIQUID_RANGE;    // we are on water surface, so next z equal liquid level
-            }
-
-            isInLiquid = true;
-
-            float ground = nextPos.z;
-            if (unitTarget->GetMap()->GetHeightInRange(unitTarget->GetPhaseMask(), nextPos.x, nextPos.y, ground))
-            {
-                if (nextPos.z < ground)
-                {
-                    nextPos.z = ground;
-                    isOnGround = true;                          // player is on ground of the water
-                }
-            }
-        }
-
-        //unitTarget->SummonCreature(VISUAL_WAYPOINT, nextPos.x, nextPos.y, nextPos.z, 0, TEMPSPAWN_TIMED_DESPAWN, 15000);
-        float hitZ = nextPos.z + 1.5f;
-        if (unitTarget->GetMap()->GetHitPosition(prevPos.x, prevPos.y, prevPos.z + 1.5f, nextPos.x, nextPos.y, hitZ, unitTarget->GetPhaseMask(), -1.0f))
-        {
-            //sLog.outString("Blink collision detected!");
-            nextPos = prevPos;
-            break;
-        }
-
-        if (isOnGround)
-        {
-            // project vector to get only positive value
-            float ac = fabs(prevPos.z - nextPos.z);
-
-            // compute slope (in radian)
-            float slope = atan(ac / step);
-
-            // check slope value
-            if (slope > MAX_SLOPE_IN_RADIAN)
-            {
-                //sLog.outString("bad slope detected! %4.2f max %4.2f, ac(%4.2f)", slope * 180 / M_PI_F, maxSlope, ac);
-                nextPos = prevPos;
-                break;
-            }
-            //sLog.outString("slope is ok! %4.2f max %4.2f, ac(%4.2f)", slope * 180 / M_PI_F, maxSlope, ac);
-        }
-
-        //sLog.outString("point %u is ok, coords %s", i, nextPos.toString().c_str());
-        nextZPointEstimation = (nextPos.z - prevPos.z) / 2.0f;
-        isPrevInLiquid = isInLiquid;
-        prevPos = nextPos;
-    }
-
-    unitTarget->NearTeleportTo(nextPos.x, nextPos.y, nextPos.z, orientation, unitTarget == m_caster);
-}
-
-void Spell::EffectLeapBack(SpellEffectEntry const* effect)
-{
-    if (unitTarget->IsTaxiFlying())
-    {
-        return;
-    }
-
-    m_caster->KnockBackFrom(unitTarget, float(effect->EffectMiscValue) / 10, float(damage) / 10);
-}
-
-/**
- * @brief Modifies player reputation for the faction referenced by the effect.
- *
- * @param effect The effect index containing faction and reputation values.
- */
-void Spell::EffectReputation(SpellEffectEntry const* effect)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Player* _player = (Player*)unitTarget;
-
-    int32  rep_change = m_currentBasePoints[effect->EffectIndex];
-    uint32 faction_id = effect->EffectMiscValue;
-
-    FactionEntry const* factionEntry = sFactionStore.LookupEntry(faction_id);
-
-    if (!factionEntry)
-    {
-        return;
-    }
-
-    rep_change = _player->CalculateReputationGain(REPUTATION_SOURCE_SPELL, rep_change, faction_id);
-
-    _player->GetReputationMgr().ModifyReputation(factionEntry, rep_change);
-}
-
-/**
- * @brief Marks the referenced quest objective as completed for the player target.
- *
- * @param effect The effect index containing the quest id.
- */
-void Spell::EffectQuestComplete(SpellEffectEntry const* effect)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    // A few spells has additional value from basepoints, check condition here.
-    switch (m_spellInfo->Id)
-    {
-        case 43458:                                         // Secrets of Nifflevar
-        {
-            if (!unitTarget->HasAura(effect->CalculateSimpleValue()))
-            {
-                return;
-            }
-
-            break;
-        }
-        // TODO: implement these!
-        // "this spell awards credit for the entire raid (all spell targets as this is area target) if just ONE member has both auras (yes, both effect's basepoints)"
-        // case 72155:                                      // Harvest Blight Specimen
-        // case 72162:                                      // Harvest Blight Specimen
-        // break;
-        default:
-            break;
-    }
-
-    uint32 quest_id = effect->EffectMiscValue;
-    ((Player*)unitTarget)->AreaExploredOrEventHappens(quest_id);
-}
-
-/**
- * @brief Resurrects the target player with flat or percentage-based health and mana.
- *
- * @param effect The effect index containing resurrection resource data.
- */
-void Spell::EffectSelfResurrect(SpellEffectEntry const* effect)
-{
-    if (!unitTarget || unitTarget->IsAlive())
-    {
-        return;
-    }
-    if (unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-    if (!unitTarget->IsInWorld())
-    {
-        return;
-    }
-
-    uint32 health = 0;
-    uint32 mana = 0;
-
-    // flat case
-    if (damage < 0)
-    {
-        health = uint32(-damage);
-        mana = effect->EffectMiscValue;
-    }
-    // percent case
-    else
-    {
-        health = uint32(damage / 100.0f * unitTarget->GetMaxHealth());
-        if (unitTarget->GetMaxPower(POWER_MANA) > 0)
-        {
-            mana = uint32(damage / 100.0f * unitTarget->GetMaxPower(POWER_MANA));
-        }
-    }
-
-    Player* plr = ((Player*)unitTarget);
-    plr->ResurrectPlayer(0.0f);
-
-    plr->SetHealth(health);
-    plr->SetPower(POWER_MANA, mana);
-    plr->SetPower(POWER_RAGE, 0);
-    plr->SetPower(POWER_ENERGY, plr->GetMaxPower(POWER_ENERGY));
-
-    plr->SpawnCorpseBones();
-}
-
-/**
- * @brief Opens skinning loot for a creature and updates the player's gathering skill.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectSkinning(SpellEffectEntry const* /*effect*/)
-{
-    if (unitTarget->GetTypeId() != TYPEID_UNIT)
-    {
-        return;
-    }
-    if (!m_caster || m_caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Creature* creature = (Creature*) unitTarget;
-    int32 targetLevel = creature->getLevel();
-
-    uint32 skill = creature->GetCreatureInfo()->GetRequiredLootSkill();
-
-    ((Player*)m_caster)->SendLoot(creature->GetObjectGuid(), LOOT_SKINNING);
-    creature->RemoveFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_SKINNABLE);
-
-    int32 reqValue = targetLevel < 10 ? 0 : targetLevel < 20 ? (targetLevel - 10) * 10 : targetLevel * 5;
-
-    int32 skillValue = ((Player*)m_caster)->GetPureSkillValue(skill);
-
-    // Double chances for elites
-    ((Player*)m_caster)->UpdateGatherSkill(skill, skillValue, reqValue, creature->IsElite() ? 2 : 1);
-}
-
-/**
- * @brief Moves the caster into melee contact with the target and starts attacking if appropriate.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectCharge(SpellEffectEntry const* /*effect*/)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-
-    // TODO: research more ContactPoint/attack distance.
-    // 3.666666 instead of ATTACK_DISTANCE(5.0f) in below seem to give more accurate result.
-    float x, y, z;
-    unitTarget->GetContactPoint(m_caster, x, y, z, 3.666666f);
-
-    if (unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        ((Creature*)unitTarget)->StopMoving();
-    }
-
-    // Only send MOVEMENTFLAG_WALK_MODE, client has strange issues with other move flags
-    m_caster->MonsterMoveWithSpeed(x, y, z, 24.f, true, true);
-
-    // not all charge effects used in negative spells
-    if (unitTarget != m_caster && !IsPositiveSpell(m_spellInfo->Id))
-    {
-        m_caster->Attack(unitTarget, true);
-    }
-}
-
-void Spell::EffectCharge2(SpellEffectEntry const* /*effect*/)
-{
-    float x, y, z;
-    if (m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION)
-    {
-        m_targets.getDestination(x, y, z);
-
-        if (unitTarget->GetTypeId() != TYPEID_PLAYER)
-        {
-            ((Creature*)unitTarget)->StopMoving();
-        }
-    }
-    else if (unitTarget && unitTarget != m_caster)
-    {
-        unitTarget->GetContactPoint(m_caster, x, y, z, 3.666666f);
-    }
-    else
-    {
-        return;
-    }
-
-    // Only send MOVEMENTFLAG_WALK_MODE, client has strange issues with other move flags
-    m_caster->MonsterMoveWithSpeed(x, y, z, 24.f, true, true);
-
-    // not all charge effects used in negative spells
-    if (unitTarget && unitTarget != m_caster && !IsPositiveSpell(m_spellInfo->Id))
-    {
-        m_caster->Attack(unitTarget, true);
-    }
-}
-
-/**
- * @brief Applies a knockback to a player target.
- *
- * @param effect The effect index containing horizontal speed data.
- */
-void Spell::EffectKnockBack(SpellEffectEntry const* effect)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-
-    unitTarget->KnockBackFrom(m_caster, float(effect->EffectMiscValue) / 10, float(damage) / 10);
-}
-
-/**
- * @brief Starts the taxi path referenced by the spell effect for a player target.
- *
- * @param effect The effect index containing the taxi path id.
- */
-void Spell::EffectSendTaxi(SpellEffectEntry const* effect)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    ((Player*)unitTarget)->ActivateTaxiPathTo(effect->EffectMiscValue, m_spellInfo->Id);
-}
-
-/**
- * @brief Pulls a player target toward the caster using reverse knockback.
- *
- * @param effect The effect index containing vertical speed data.
- */
-void Spell::EffectPlayerPull(SpellEffectEntry const* effect)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-
-    float x, y, z;
-    m_caster->GetPosition(x, y, z);
-
-    // move back a bit
-    x = x - (0.6 * cos(m_caster->GetOrientation() + M_PI_F));
-    y = y - (0.6 * sin(m_caster->GetOrientation() + M_PI_F));
-
-    // Try to normalize Z coord because GetContactPoint do nothing with Z axis
-    unitTarget->UpdateAllowedPositionZ(x, y, z);
-
-    float speed = m_spellInfo->speed ? m_spellInfo->speed : 27.0f;
-    unitTarget->GetMotionMaster()->MoveJump(x, y, z, speed, 2.5f);
-}
